10_flashes/requirements.txt
flask
flask-bootstrap
flask-sqlalchemy
flask-wtf
python-dotenv
email_validator
